[Patchset 16] Log histogram of unwind timing.
Test: profile, get go/heapprofd-example-histogram as output.

Bug: 129054982

+++ src/profiling/memory/heapprofd_producer.h
@@ -46,7 +46,7 @@ struct Process {
 
 class LogHistogram {
  public:
-  static const size_t kMaxBucket;
+  static const uint64_t kMaxBucket;
   static constexpr size_t kBuckets = 20;
 
   void Add(uint64_t value) { values_[GetBucket(value)]++; }
